+++ src/App.tsx
@@ -1,3 +1,16 @@
-export default function App() {
-  return <>我是一个页面</>;
+import 'antd/dist/reset.css';
+import { Provider } from 'react-redux';
+import { ConfigProvider } from 'antd';
+
+function AppConfig() {
+  return <ConfigProvider></ConfigProvider>;
 }
+
+function App() {
+  return (
+    <Provider>
+      <AppConfig />
+    </Provider>
+  );
+}
+export default App;