+++ src/layouts/index.tsx
@@ -1,0 +1,3 @@
+// https://gitee.com/kong_yiji_and_lavmi/react-ant-admin/blob/vite/src/router/list.tsx
+//https://github.com/lichenbuliren/fe-project-base/blob/main/src/routes/index.ts 
+// https://github.com/JS-banana/vite-react-ts/blob/master/src/pages/Redirect.tsx