+++ src/routes/staticRoutes.tsx
@@ -1,0 +1,11 @@
+import { RouteConfig } from 'react-router-config';
+
+const staticRoutes: RouteConfig[] = [
+  {
+    path: '/',
+    exact: true,
+    // component: Home,
+  },
+];
+
+export default staticRoutes;